requirements.txt
# Semantic-Cluster-WebUI Requirements
# Phase-0

# ============================================
# Core Dependencies
# ============================================

# Configuration Management
omegaconf>=2.3.0
pyyaml>=6.0

# Web UI Framework
gradio>=4.0.0

# ============================================
# Phase-1: MVP Dependencies
# ============================================

# Deep Learning & Computer Vision
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Data Processing
numpy>=1.24.0,<2
pandas>=2.0.0
numexpr>=2.8.4
bottleneck>=1.3.6

# Clustering
hdbscan>=0.8.33

# ============================================
# Phase-2: VLM Dependencies（Step-4/5 图像描述与簇标签）
# ============================================

# Vision-Language Models（Qwen2-VL 需 4.37+）
transformers>=4.37.0
accelerate>=0.24.0
sentencepiece>=0.1.99
huggingface_hub>=0.20.0
jinja2>=3.0

# ============================================
# Phase-5: Visualization Dependencies
# ============================================

# Dimensionality Reduction
umap-learn>=0.5.4

# Interactive Plotting
plotly>=5.17.0

# Static Plotting (for scripts)
matplotlib>=3.7.0

# ============================================
# Phase-7: Optional Performance Dependencies
# ============================================

# GPU Acceleration (Optional)
# cuml>=23.0.0  # Requires CUDA
# faiss-gpu>=1.7.4  # Requires CUDA

# Fast Inference (Optional)
# vllm>=0.2.0  # For VLM acceleration

# VLM 加速（可选，需 CUDA 且与 transformers 版本兼容）
# pip install flash-attn --no-build-isolation
# config: vlm.use_flash_attn: true

# ============================================
# Utilities
# ============================================

# Progress Bars
tqdm>=4.66.0

# File Management
pathlib>=1.0.1

# Logging
loguru>=0.7.0

# ============================================
# Development Dependencies (Optional)
# ============================================

# Code Formatting
# black>=23.0.0
# isort>=5.12.0

# Testing
# pytest>=7.4.0
# pytest-cov>=4.1.0
